+++ lib/persistor/file_persistor/extensions/document.dart
@@ -32,13 +32,13 @@ extension DocumentExtensions<T> on Document<T> {
     return false;
   }
 
-  FilePersistDocument<T> toPersistenceDoc() {
-    return FilePersistDocument<T>(
+  FilePersistDocument toPersistenceDoc() {
+    return FilePersistDocument(
       id: id,
       parent: parent,
       encrypted: isEncrypted(),
       key: getPersistenceKey(),
-      data: getJson(),
+      data: getSerialized(),
     );
   }
 }